extract object aliases to a trait
